flaskapp/requirements.txt
pillow
Flask
gunicorn
tensorflow
numpy